--- a/src/imp.rs
+++ b/src/imp.rs
@@ -1,5 +1,5 @@
 use std::{
-    any::type_name,
+    any::{type_name, TypeId},
     hash::BuildHasherDefault,
     os::raw::c_int,
     sync::atomic::{AtomicBool, AtomicUsize, Ordering::Relaxed},
@@ -13,7 +13,7 @@ use crate::{AllCounts, Counts};
 
 static ENABLE: AtomicBool = AtomicBool::new(cfg!(feature = "print_at_exit"));
 
-type GlobalStore = DashMap<&'static str, Store, BuildHasherDefault<FxHasher>>;
+type GlobalStore = DashMap<TypeId, Store, BuildHasherDefault<FxHasher>>;
 
 #[inline]
 fn global_store() -> &'static GlobalStore {
@@ -45,39 +45,44 @@ fn enabled() -> bool {
 }
 
 #[inline]
-pub(crate) fn dec<T>() {
+pub(crate) fn dec<T: 'static>() {
     if enabled() {
-        do_dec(type_name::<T>())
+        do_dec(TypeId::of::<T>())
     }
 }
 #[inline(never)]
-fn do_dec(key: &'static str) {
+fn do_dec(key: TypeId) {
     if let Some(store) = global_store().get(&key) {
         store.value().dec();
     }
 }
 
 #[inline]
-pub(crate) fn inc<T>() {
+pub(crate) fn inc<T: 'static>() {
     if enabled() {
-        do_inc(type_name::<T>())
+        do_inc(TypeId::of::<T>(), type_name::<T>())
     }
 }
 #[inline(never)]
-fn do_inc(key: &'static str) {
-    global_store().entry(&key).or_default().value().inc();
+fn do_inc(key: TypeId, name: &'static str) {
+    global_store().entry(key).or_insert_with(|| Store { name, ..Store::default() }).value().inc();
 }
 
-pub(crate) fn get<T>() -> Counts {
-    do_get(type_name::<T>())
+pub(crate) fn get<T: 'static>() -> Counts {
+    do_get(TypeId::of::<T>())
 }
-fn do_get(key: &'static str) -> Counts {
-    global_store().entry(&key).or_default().value().read()
+fn do_get(key: TypeId) -> Counts {
+    global_store().entry(key).or_default().value().read()
 }
 
 pub(crate) fn get_all() -> AllCounts {
-    let mut entries =
-        global_store().iter().map(|entry| (*entry.key(), entry.value().read())).collect::<Vec<_>>();
+    let mut entries = global_store()
+        .iter()
+        .map(|entry| {
+            let store = entry.value();
+            (store.type_name(), store.read())
+        })
+        .collect::<Vec<_>>();
     entries.sort_by_key(|(name, _counts)| *name);
     AllCounts { entries }
 }
@@ -87,6 +92,7 @@ struct Store {
     total: AtomicUsize,
     max_live: AtomicUsize,
     live: AtomicUsize,
+    name: &'static str,
 }
 
 impl Store {
@@ -106,5 +112,9 @@ impl Store {
             max_live: self.max_live.load(Relaxed),
             live: self.live.load(Relaxed),
         }
+    }
+
+    fn type_name(&self) -> &'static str {
+        self.name
     }
 }