@matklad 2021-11-24
Could you bump cargo toml version here as well, so that this gets a release automatically?