@WaffleLapkin 2021-11-24
Sure, I've updated version to `3.0.0`